+++ qt-models/filtermodels.h
@@ -7,64 +7,62 @@
 #include <stdint.h>
 #include <vector>
 
-class MultiFilterInterface {
+class FilterModelBase : public QStringListModel {
+	Q_OBJECT
 public:
-	MultiFilterInterface() : anyChecked(false) {}
 	virtual bool doFilter(struct dive *d, QModelIndex &index0, QAbstractItemModel *sourceModel) const = 0;
-	virtual void clearFilter() = 0;
+	void clearFilter();
+	void selectAll();
+	void invertSelection();
 	std::vector<char> checkState;
 	bool anyChecked;
-};
-
-class FilterModelBase : public QStringListModel, public MultiFilterInterface {
+	bool negate;
+public
+slots:
+	void setNegate(bool negate);
 protected:
 	explicit FilterModelBase(QObject *parent = 0);
 	void updateList(const QStringList &new_list);
+	virtual int countDives(const char *) const = 0;
+private:
+	Qt::ItemFlags flags(const QModelIndex &index) const;
+	QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
+	bool setData(const QModelIndex &index, const QVariant &value, int role = Qt::EditRole);
 };
 
 class TagFilterModel : public FilterModelBase {
 	Q_OBJECT
 public:
 	static TagFilterModel *instance();
-	virtual QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
-	virtual bool setData(const QModelIndex &index, const QVariant &value, int role = Qt::EditRole);
-	virtual Qt::ItemFlags flags(const QModelIndex &index) const;
 	bool doFilter(struct dive *d, QModelIndex &index0, QAbstractItemModel *sourceModel) const;
-	void clearFilter();
 public
 slots:
 	void repopulate();
 
 private:
 	explicit TagFilterModel(QObject *parent = 0);
+	int countDives(const char *) const;
 };
 
 class BuddyFilterModel : public FilterModelBase {
 	Q_OBJECT
 public:
 	static BuddyFilterModel *instance();
-	virtual QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
-	virtual bool setData(const QModelIndex &index, const QVariant &value, int role = Qt::EditRole);
-	virtual Qt::ItemFlags flags(const QModelIndex &index) const;
 	bool doFilter(struct dive *d, QModelIndex &index0, QAbstractItemModel *sourceModel) const;
-	void clearFilter();
 public
 slots:
 	void repopulate();
 
 private:
 	explicit BuddyFilterModel(QObject *parent = 0);
+	int countDives(const char *) const;
 };
 
 class LocationFilterModel : public FilterModelBase {
 	Q_OBJECT
 public:
 	static LocationFilterModel *instance();
-	virtual QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
-	virtual bool setData(const QModelIndex &index, const QVariant &value, int role = Qt::EditRole);
-	virtual Qt::ItemFlags flags(const QModelIndex &index) const;
 	bool doFilter(struct dive *d, QModelIndex &index0, QAbstractItemModel *sourceModel) const;
-	void clearFilter();
 public
 slots:
 	void repopulate();
@@ -73,23 +71,21 @@ slots:
 
 private:
 	explicit LocationFilterModel(QObject *parent = 0);
+	int countDives(const char *) const;
 };
 
 class SuitsFilterModel : public FilterModelBase {
 	Q_OBJECT
 public:
 	static SuitsFilterModel *instance();
-	virtual QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
-	virtual bool setData(const QModelIndex &index, const QVariant &value, int role = Qt::EditRole);
-	virtual Qt::ItemFlags flags(const QModelIndex &index) const;
 	bool doFilter(struct dive *d, QModelIndex &index0, QAbstractItemModel *sourceModel) const;
-	void clearFilter();
 public
 slots:
 	void repopulate();
 
 private:
 	explicit SuitsFilterModel(QObject *parent = 0);
+	int countDives(const char *) const;
 };
 
 class MultiFilterSortModel : public QSortFilterProxyModel {
@@ -97,8 +93,8 @@ class MultiFilterSortModel : public QSortFilterProxyModel {
 public:
 	static MultiFilterSortModel *instance();
 	virtual bool filterAcceptsRow(int source_row, const QModelIndex &source_parent) const;
-	void addFilterModel(MultiFilterInterface *model);
-	void removeFilterModel(MultiFilterInterface *model);
+	void addFilterModel(FilterModelBase *model);
+	void removeFilterModel(FilterModelBase *model);
 	int divesDisplayed;
 public
 slots:
@@ -111,7 +107,7 @@ signals:
 	void filterFinished();
 private:
 	MultiFilterSortModel(QObject *parent = 0);
-	QList<MultiFilterInterface *> models;
+	QList<FilterModelBase *> models;
 	bool justCleared;
 	struct dive_site *curr_dive_site;
 };