+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- desktop: Add negation button to filter lists (#735)
+- desktop: Add select all/select none/invert selection context menu to filter lists
 - Fix dive site merging for cloud (and local git storage) users. (#939)
 - desktop: Add variable zoom-levels for the dive photos tab (#898)
 - mobile: Faulty navigation after aborted dive edit (#932)